# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PR Newsroom API Documentation",
        "description": "Made with ❤ by Pikiran Rakyat.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "newsroom@pikiran-rakyat.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
        "url": "/api/",
        "description": "Production Server"
        },
        {
        "url": "/",
        "description": "Local Server"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "",
                "description": "Get all data",
                "operationId": "user_get_all",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per_page value is number. ex : ?per_page=10",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page value is number. ex : ?page=2",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort value is string with rule column-name:order. ex : ?sort=id:asc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "where",
                        "in": "query",
                        "description": "Where value is object. ex : ?where={'name':'john', 'dob':'1990-12-31'}",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "where",
                        "in": "query",
                        "description": "Where value is object. ex : ?where={'name':'john', 'dob':'1990-12-31'}",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Count value is boolean. ex : ?count=true",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "",
                "description": "Insert data",
                "operationId": "user_insert",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Insert Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "",
                "description": "Get data by id",
                "operationId": "user_get_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "",
                "description": "Update data",
                "operationId": "user_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Update Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "",
                "description": "Delete data",
                "operationId": "user_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Delete Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user_level": {
            "get": {
                "tags": [
                    "User Level"
                ],
                "summary": "",
                "description": "Get all data",
                "operationId": "user_get_all",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per_page value is number. ex : ?per_page=10",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page value is number. ex : ?page=2",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort value is string with rule column-name:order. ex : ?sort=id:asc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "where",
                        "in": "query",
                        "description": "Where value is object. ex : ?where={'name':'john', 'dob':'1990-12-31'}",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "where",
                        "in": "query",
                        "description": "Where value is object. ex : ?where={'name':'john', 'dob':'1990-12-31'}",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Count value is boolean. ex : ?count=true",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Get Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Level"
                ],
                "summary": "",
                "description": "Insert data",
                "operationId": "user_insert",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "success": true,
                                    "message": "Insert Data Successfull",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}